==> scripts/01_prepare_brightway.ipynb <==
# 01: Prepare brightway

In this script, we create a new brightway project, import ecoinvent and add all necessary LCIA methods.

%run common_definitions.py

import brightway2 as bw
from premise_gwp import add_premise_gwp

import pandas as pd
import os

## Create a new brightway project

bw.projects.set_current(BW_PROJECTNAME)

## Create output folder

output_fp = "../output/" + BW_PROJECTNAME
if not os.path.exists(output_fp):
    os.mkdir(output_fp)

## Import ecoinvent

ei_path = "C:/Users/davidba/LCA/ecoinvent/ecoinvent {}_cutoff_ecoSpold02/datasets/".format(EI_VERSION)

if "biosphere3" not in bw.databases:
    print("Setting up missing biosphere.")
    bw.bw2setup()
ei_label = "ecoinvent {} cutoff".format(EI_VERSION)
if ei_label not in bw.databases:
    print("Importing ecoinvent DB.")
    ei = bw.SingleOutputEcospold2Importer(
        ei_path, ei_label, use_mp=False)
    ei.apply_strategies()
    ei.statistics()
    ei.write_database()

## Adding premise_gwp methods

This function adds climate change LCIA methods with biogenic flows.

add_premise_gwp()

## Create split climate change LCIA methods

Now the newly created climate change LCIA methods are split into the sources `biogenic`, `land use`, and `fossil`.

base_methods = [
    ('IPCC 2021', 'climate change', 'GWP 100a, incl. H and bio CO2'),
    ('IPCC 2021', 'climate change', 'GWP 20a, incl. H and bio CO2')
]

for method in base_methods:
    M = bw.Method(method)
    unit = ""
    if "unit" in M.metadata:
        unit = M.metadata["unit"]

    CFs_bio = []
    CFs_landuse = []
    CFs_fossil = []
    for flow in bw.Method(method).load():
        act = bw.get_activity(flow[0])
        if "non-fossil" in act["name"] or "in air" in act["name"]:
            CFs_bio.append(list(flow))
        elif "soil or biomass" in act["name"]:
            CFs_landuse.append(list(flow))
        else:
            CFs_fossil.append(list(flow))

    new_key = (method[0], method[1]+": biogenic", method[2].split(",")[0])    
    new_method = bw.Method(new_key)
    new_method.validate(CFs_bio)
    new_method.register(unit=unit)
    new_method.write(CFs_bio)

    new_key = (method[0], method[1]+": land use", method[2].split(",")[0])    
    new_method = bw.Method(new_key)
    new_method.validate(CFs_landuse)
    new_method.register(unit=unit)
    new_method.write(CFs_landuse)

    new_key = (method[0], method[1]+": fossil", method[2].split(",")[0])    
    new_method = bw.Method(new_key)
    new_method.validate(CFs_fossil)
    new_method.register(unit=unit)
    new_method.write(CFs_fossil)

## Add the Ecocosts Materials method

table = pd.read_csv("../data/cfs_ecocosts_metals.csv")
table.dropna(subset="characterization factor", inplace=True)

db = bw.Database("biosphere3")
resources = [flow for flow in db if flow["categories"] == ("natural resource", "in ground")]

new_key = ("Eco-costs", "Material scarcity", "VAR(95)")
CFs = []
for idx, row in table.iterrows():
    matches = [flow for flow in resources if flow["name"] == row["name"] and str(flow["categories"]) == row["categories"]]
    if len(matches) == 1:
        CFs.append([matches[0].key, row["characterization factor"]])
    else:
        print("Error: More than one match.")

new_method = bw.Method(new_key)
new_method.validate(CFs)
new_method.register(unit="EUR2023")
new_method.write(CFs)

## Add the old ReCiPe methods

Load CFs.

df_recipe = pd.read_csv("../data/cfs_old_recipe_methods_mapped.csv")

Find matching biosphere flows

db = bw.Database("biosphere3")

flowlist = []
categorieslist = []
keylist = []

needed_flows = list(df_recipe.groupby(["new flow name", "flow categories"]).size().index)

for j, f in enumerate(needed_flows):
    matching_names = [act for act in db if act["name"] == f[0]]
    matching_cats = [act for act in matching_names if str(act["categories"]) == f[1]]
    if len(matching_cats) == 1:
        flowlist.append(f[0])
        categorieslist.append(f[1])
        keylist.append(matching_cats[0].key)
    print("{} of {} flows checked.".format(j+1, len(needed_flows)))

flow_mapping = pd.DataFrame(
    {
        "new flow name": flowlist,
        "flow categories": categorieslist,
        "mapped key": keylist
    }
)

df_recipe_mapped = pd.concat(
    (
        df_recipe.set_index(["new flow name", "flow categories"]),
        flow_mapping.set_index(["new flow name", "flow categories"])
    ),
    axis=1
).reset_index()

df_recipe_mapped[["method", "unit", "new flow name", "flow categories", "mapped key", "CF"]].to_csv("../data/cfs_old_recipe_methods_mapped_w_keys.csv", index=False)

Add the methods to brightway:

recipe_cfs = pd.read_csv("../data/cfs_old_recipe_methods_mapped_w_keys.csv").dropna(subset="mapped key")
methods2add = list(recipe_cfs["method"].unique())

for m in methods2add:
    sel = recipe_cfs[recipe_cfs["method"] == m].copy()
    new_key = tuple(list(m.split(", ")))
    unit = sel["unit"].iloc[0]
    keys = list(sel["mapped key"])
    cfs = list(sel["CF"])
    new_cfs = [[eval(k), cf] for k, cf in zip(keys, cfs)]
    new_method = bw.Method(new_key)
    new_method.validate(new_cfs)
    new_method.register(unit=unit)
    new_method.write(new_cfs)

    print("Added method {}".format(str(new_method)))
