==> scripts/02_run_premise.ipynb <==
# 02: Run premise

In this script, we run premise to create new prospective life cycle inventories for selected scenarios and years.

%run common_definitions.py

from premise import NewDatabase
from premise.utils import eidb_label

import time
import gc

import brightway2 as bw

bw.projects.set_current(BW_PROJECTNAME)

RUNS_PATH = "../data"

def get_shorter_version_number(ei_version):
    splits = ei_version.split(".")
    if len(splits) == 2:
        return ei_version
    if len(splits) > 2:
        return ".".join(splits[0:2])

def create_inventories(scenario, years, ei_version, update_pers_transport=False):
    """
    Create REMIND-ecoinvent inventories for given years and
    scenario. Inventories for scenarios that are to be
    omitted are deleted from the database to speed up
    the database.
    """
    ei_label = "ecoinvent {} cutoff".format(ei_version)

    print("Creating inventories for {}.".format(scenario))

    dblist = [db for db in bw.databases]

    for year in years:

        scen = {"model": "remind", "pathway": scenario, "year": year, "filepath": RUNS_PATH}
        dbname = eidb_label(scen, ei_version)
        if dbname in dblist:
            print("Existing database found for {}, year {}.".format(scenario, year))
            continue
        
        start_time = time.time()
        ndb = NewDatabase(scenarios=[scen],
                            source_db=ei_label,
                            source_version=ei_version,
                            keep_uncertainty_data=True
                            # key="tUePmX_S5B8ieZkkM7WUU2CnO8SmShwmAeWK9x2rTFo=")
        )
        ndb.update()
        if update_pers_transport:
            ndb.update("buses")
            ndb.update("two_wheelers")
            ndb.update("cars")
        ndb.write_db_to_brightway()
        print("Creation of scenario {} took: {:.1f} min"
                .format(scenario, (time.time() - start_time)/60.))
        del ndb
        gc.collect()

for scen in SCENARIOS:
    create_inventories(scen, YEARS, EI_VERSION)

bw.databases
