==> scripts/05b_calculate_combustion_impacts.ipynb <==
# 05b: Calculate combustion impacts

For liquid fuels, we also calculate the impacts from combustion to later add to the production impacts.

%run common_definitions.py

import pandas as pd
import numpy as np
import xarray as xr

import brightway2 as bw

bw.projects.set_current(BW_PROJECTNAME)
output_fp = "../output/" + BW_PROJECTNAME

### Get lookup dictionary for database codes of biosphere flows

def category_tuple_to_str(tpl):
    return "::".join(list(tpl))

DF_BASE = pd.read_csv("../data/liquids_combustion_efs_base.csv")
SCALING_FACTORS = pd.read_csv("../data/liquids_combustion_scaling_factors.csv")
BS = bw.Database("biosphere3")

flow2code = DF_BASE.groupby(["name", "categories"]).count().reset_index()[["name", "categories"]]
nonfossil_flows = flow2code[flow2code["name"].str.contains("fossil")].copy()
nonfossil_flows["name"] = nonfossil_flows["name"].str.replace("fossil", "non-fossil")
flow2code = pd.concat((flow2code, nonfossil_flows), axis=0)
matches = []
for name, cat in zip(flow2code["name"], flow2code["categories"]):
        matches_name = [flow for flow in BS if flow["name"] == name]
        matches_cat = [flow for flow in matches_name if category_tuple_to_str(flow["categories"]) == cat]
        if len(matches_cat) == 1:
            matches.append(matches_cat[0])
        else:
            raise ValueError("Not exactly one matching flow found!")

flow2code["code"] = [flow["code"] for flow in matches]

flow2code = flow2code.set_index(["name", "categories"])["code"]

### Create full dataframe of emission factors from base fuels

def get_biosphere_exchanges(base_fuel, scaling_fuel=None, is_non_fossil=False):
    sel = DF_BASE[DF_BASE["fuel name"] == base_fuel][["name", "categories", "amount"]]

    if scaling_fuel is not None:
        scaling = SCALING_FACTORS[SCALING_FACTORS["fuel name"] == scaling_fuel][["name", "categories", "scaling factor"]].fillna(1.0)
        df = pd.concat(
            (sel.set_index(["name", "categories"]),
             scaling.set_index(["name", "categories"])),
             axis=1
        )
        df["amount"] = df["amount"] * df["scaling factor"]
        df = df["amount"].reset_index()
    else:
        df = sel

    if is_non_fossil:
        df["name"] = df["name"].str.replace("fossil", "non-fossil")

    # get codes from lookup
    df["code"] = [flow2code.loc[(n, c)] for n, c in zip(df["name"], df["categories"])]

    return df

def check_fossil(row):
    tech = row["short name"]
    return ("bio" in tech or "electrolysis" in tech)

def get_scaling_fuel(row):
    val = row["scaling_fuel"]
    if isinstance(val, str):
        return val
    else:
        return None
    
def get_suffix(row):
    return ", burned in " + row["base_fuel"].split(", ")[-1]

def get_ncv(row):
    prod = row["reference product"]

    df_ncvs = pd.read_csv("../data/NCVs_v2.csv")
    return df_ncvs.set_index("product name").loc[prod]["NCV in MJ/product"]

mapping = pd.read_csv("../mappings/liquids_2_combustion_exchanges.csv")
long2short = dict(zip(list(mapping["ecoinvent name"]), list(mapping["short name"])))

dflist = []
for idx, row in mapping.iterrows():
    ncv = get_ncv(row)
    df_bs = get_biosphere_exchanges(row["base_fuel"],
                                    scaling_fuel=get_scaling_fuel(row),
                                    is_non_fossil=check_fossil(row))
    df_bs["ecoinvent name"] = row["ecoinvent name"]
    dflist.append(df_bs)

all_efs = pd.concat(dflist, axis=0)
all_efs.to_csv("../data/liquids_combustion_all_efs.csv", index=False)

### Calculate the impacts for the needed LCIA methods

def get_cfs(keys, methods):
    # initialize values
    cfs = np.zeros((len(keys), len(methods)))

    # loop through methods
    for j, m in enumerate(methods):
        print("Collecting CFs for method {} of {}, {}".format(j+1, len(methods), str(m)))
        for flow in bw.Method(m).load():
            key = flow[0][1]
            if key in keys:
                i = keys.index(key)
                cfs[i,j] = flow[1]

    # construct array
    da = xr.DataArray(
        cfs,
        coords={
            "code": keys,
            "LCIA method": [", ".join(list(m)) for m in methods]
        }
    )

    return da

df2 = pd.read_csv("../data/mfs_all_factors.csv").dropna(subset="LCIA method")

needed_methods = list(df2["LCIA method"].unique())
methods = [m for m in list(bw.methods) if ", ".join(list(m)) in needed_methods]

cc_methods = [
    ("IPCC 2021", "climate change: biogenic", "GWP 100a"),
    ("IPCC 2021", "climate change: land use", "GWP 100a"),
    ("IPCC 2021", "climate change: fossil", "GWP 100a")
]
methods = methods + cc_methods

print("{} methods needed, {} methods found".format(len(needed_methods)+len(cc_methods), len(methods)))

unique_keys = list(all_efs["code"].unique())

cfs = get_cfs(unique_keys, methods)

### Build emission factor data array

ef_xr = all_efs.set_index(["ecoinvent name", "code"])["amount"].to_xarray().fillna(0)

### Calculate impacts

Impacts are simply the (matrix) product of characterization and emission factors.

impacts = (ef_xr * cfs).sum(dim="code")
impact_df = impacts.to_pandas()

impact_df.to_csv(output_fp+"/impacts_liquids_combustion.csv")
