==> scripts/07b_extended_sensitivity_analysis.ipynb <==
# 07b: Extended sensitivity analysis

In this script, we perform sensitivity analysis by sampling technosphere, biosphere, and weighting (monetization) inputs.

from lca_sensitivity_analysis import *

import xarray as xr
from scipy import sparse
import pandas as pd
from time import time

import seaborn as sns
import matplotlib.pyplot as plt

import multiprocessing as mp

%run common_definitions.py

bw.projects.set_current(BW_PROJECTNAME)
output_fp = "../output/" + BW_PROJECTNAME

### Select scenario and year and further parameter choices

scen = "SSP2-NPi"
year = 2030
substring = "{}_{}".format(scen, str(year))
db = bw.Database([name for name in bw.databases if substring in name][0])

### Select technologies

df = pd.read_csv("../mappings/technology_selection_energy_provision_v4.csv")
long2short = dict(zip(list(df["ecoinvent name"]), list(df["short name"])))

subset_names = list(df["ecoinvent name"].unique())
db_subset = [act for act in db if act["name"] in subset_names]

def contains_global(actlist):
    locations = [act["location"] for act in actlist]

    return any(x in locations for x in ["RoW", "GLO", "World"])

def select_region(actlist):
    locations = [act["location"] for act in actlist]

    if contains_global(actlist):
        if "GLO" in locations:
            return actlist[locations.index("GLO")]
        elif "World" in locations:
            return actlist[locations.index("World")]
        else:
            return actlist[locations.index("RoW")]
    else:
        # choose alphabetically first location (need a deterministic selection)
        first_location = sorted(locations)[0]
        return actlist[locations.index(first_location)]
    


activities = {}

for sector in df["sector"].unique():
    sel = df[df["sector"] == sector]
    all_matches = []
    for name, product in zip(list(sel["ecoinvent name"]), list(sel["reference product"])):
        matches = [act for act in db_subset if act["name"] == name and act["reference product"] == product]
        if len(matches) == 0:
            print("No matches for {}, {}".format(name, product))
        sel = select_region(matches)
        # if not contains_global(matches):
        #     print("No global dataset for {}, {}. Location {} selected.".format(name, product, sel["location"]))
        # print("\t{} activities found for {}, {} in database {}".format(len(matches), name, product, substring))
        all_matches.append(sel)
    print("A total of {} activities found in sector {} in database {}".format(len(all_matches), sector, substring))
    activities[sector] = all_matches

### Get methods and weights (monetization factors)

mfs_mc_sample = pd.read_csv("../data/mfs_monte_carlo_sample_euro{}_larger_nonzero.csv".format(str(EURO_REF_YEAR))).set_index(
    ["impact category", "sample index", "LCIA method"]
    )["MF"].to_xarray().fillna(0)

weights_total = mfs_mc_sample.sum(dim="impact category")

methods = [m for m in list(bw.methods) if ", ".join(list(m)) in weights_total.coords["LCIA method"].values]

ALL_WEIGHTS = weights_total.to_pandas().T.loc[[", ".join(list(m)) for m in methods]].to_numpy()
mean_weights = np.mean(ALL_WEIGHTS, axis=1)

## One-factor at a time analysis

def gsa_ofat(act, methods, mean_weights, Nsamples, print_to_console=True):
    # set up LCA
    fu = {act: 1}
    lca = bw.LCA(fu)
    lca.lci()

    # biosphere variation
    t0 = time()
    bio_rng = MCRandomNumberGenerator(lca.bio_params)
    bio_results = []
    for i in range(Nsamples):
        lca.rebuild_biosphere_matrix(bio_rng.next())

        count = len(lca.activity_dict)
        lca.inventory = lca.biosphere_matrix * \
            sparse.spdiags([lca.supply_array], [0], count, count)
        
        temp = []
        for m in methods:
            lca.switch_method(m)
            lca.lcia_calculation()
            temp.append(lca.score)

        bio_results.append(np.sum(np.array(temp) * mean_weights))
        print("{}: Biosphere: Run {} of {} done.".format(str(act), i+1, Nsamples))
    t1 = time()

    if print_to_console:
        print("{}: Biosphere variation runs done in {} seconds".format(str(act), round(t1-t0, 2)))

    # technosphere variation
    t0 = time()
    tech_rng = MCRandomNumberGenerator(lca.tech_params)
    tech_results = []
    for i in range(Nsamples):
        t2 = time()
        lca.rebuild_technosphere_matrix(tech_rng.next())

        t3 = time()

        lca.redo_lci(fu)

        t4 = time()

        temp = []
        for m in methods:
            lca.switch_method(m)
            lca.lcia_calculation()
            temp.append(lca.score)

        tech_results.append(np.sum(np.array(temp) * mean_weights))
        print("{}: Technosphere: Run {} of {} done (rebuilding: {} seconds, lci: {} seconds)".format(str(act), i+1, Nsamples, t3-t2, t4-t3))
    t1 = time()

    if print_to_console:
        print("{}: Technosphere variation runs done in {} seconds".format(str(act), t1-t0))

    return {
        "technosphere": tech_results,
        "biosphere": bio_results
    }

def gsa_ofat_multi(actlist, mean_cost_method, Nsamples):
    # set up LCA
    fu = {actlist[0]: 1}
    lca = bw.LCA(fu, mean_cost_method)
    lca.lci()
    lca.lcia()

    ## Biosphere variation ##
    # calculate supply arrays for all activities
    lca.decompose_technosphere()
    supply_arrays = []
    count = len(lca.activity_dict)
    for act in actlist:
        lca.redo_lci({act: 1})
        supply_arrays.append(lca.supply_array)

    bio_rng = MCRandomNumberGenerator(lca.bio_params)
    bio_results = np.zeros((len(actlist), Nsamples))
    for j in range(Nsamples):
        lca.rebuild_biosphere_matrix(bio_rng.next())
        for i, supply in enumerate(supply_arrays):
            lca.inventory = lca.biosphere_matrix * \
                    sparse.spdiags([supply], [0], count, count)
            lca.lcia_calculation()
            bio_results[i, j] = lca.score
            print("Biosphere variation: Activity {} of {}, run {} of {} done.".format(i+1, len(actlist), j+1, Nsamples))

    # technosphere variation
    # set up LCA again
    fu = {actlist[0]: 1}
    lca = bw.LCA(fu, mean_cost_method)
    lca.lci()
    lca.lcia()

    tech_rng = MCRandomNumberGenerator(lca.tech_params)
    tech_results = np.zeros((len(actlist), Nsamples))
    for j in range(Nsamples):
        lca.rebuild_technosphere_matrix(tech_rng.next())
        lca.decompose_technosphere()

        for i, act in enumerate(actlist):
            lca.redo_lci({act: 1})
            lca.lcia_calculation()
            tech_results[i, j] = lca.score
            print("Technosphere variation: Activity {} of {}, Run {} of {} done.".format(i+1, len(actlist), j+1, Nsamples))

    return {
        "technosphere": tech_results,
        "biosphere": bio_results
    }

cost_method = ("Monte Carlo monetization (larger sample)", "mean total costs")

all_activities = []
all_sectors = []
for sec, a in activities.items():
    all_activities += a
    all_sectors += len(a) * [sec,]

short2location = {}
for act in all_activities:
    tech = long2short[act["name"]]
    loc = act["location"]
    short2location[tech] = loc

scores_multi_all = gsa_ofat_multi(all_activities, cost_method, 2000)

dflist = []
for stage, results in scores_multi_all.items():
    for i, act in enumerate(all_activities):
        y = results[i,:]
        df = pd.DataFrame({
            "output value": y,
        })
        df["stage"] = stage
        df["sample index"] = np.arange(len(y))
        df["short name"] = long2short[act["name"]]
        df["sector"] = all_sectors[i]
        df["scenario"] = scen
        df["year"] = year
        df["location"] = act["location"]
        dflist.append(df)

output = pd.concat(dflist, axis=0, ignore_index=True)
output.to_csv("gsa_ofat_all_sample1.csv")

### Calculate costs

new_cs = {"inv": [{act: 1} for act in all_activities], "ia": methods}
bw.calculation_setups["GSA"] = new_cs

mc = bw.MultiLCA("GSA")
    

costs = np.matmul(mc.results, weights_total.to_pandas()[[", ".join(list(m)) for m in methods]].to_numpy().T)

idx = []
for act, sector in zip(all_activities, all_sectors):
    tech = long2short[act["name"]]
    idx.append(str((tech, sector)))

xr.DataArray(
    costs,
    coords={
        "activity key": idx,
        "sample index": list(range(costs.shape[1]))
    }
).to_netcdf(output_fp+"/gsa_ofat_all_costs.nc")

### Calculate variances and sensitivity scores

output = pd.read_csv(output_fp+"/gsa_ofat_all_outputsamples.csv")
output

def filter_by_quantiles(df, idx, qidx, col, low, high):
    dfnew = df.set_index(idx)
    low = df.groupby(qidx)[col].quantile(low).rename("low")
    high = df.groupby(qidx)[col].quantile(high).rename("high")

    dfnew = dfnew.join(low)
    dfnew = dfnew.join(high).reset_index()

    sel = dfnew[dfnew[col] <= dfnew["high"]]
    sel = sel[sel[col] >= sel["low"]]

    return sel

def calculate_variances(df, chunks=1):
    dfnew = df.copy()
    Nsamples = dfnew["sample index"].max() + 1
    dfnew["chunk index"] = dfnew["sample index"] // (Nsamples // chunks)
    
    return dfnew.groupby(["short name", "sector", "stage", "chunk index"]).agg({"output value": np.var})

def calculate_stds(df, chunks=1):
    dfnew = df.copy()
    Nsamples = dfnew["sample index"].max() + 1
    dfnew["chunk index"] = dfnew["sample index"] // (Nsamples // chunks)
    
    return dfnew.groupby(["short name", "sector", "stage", "chunk index"]).agg({"output value": np.std})

def normalize_column(df, col, values, new_value_name):
    idx = list(set(df.columns) - {col}- {values})
    a = df.pivot(index=idx, columns=col, values=values)

    return a.divide(a.sum(axis=1), axis=0).melt(
        ignore_index=False, value_name=new_value_name).reset_index()

def cv(x):
    return np.std(x) / np.mean(x)

def calculate_cvs(df, chunks=1):
    dfnew = df.copy()
    Nsamples = dfnew["sample index"].max() + 1
    dfnew["chunk index"] = dfnew["sample index"] // (Nsamples // chunks)
    
    return dfnew.groupby(["short name", "sector", "stage", "chunk index"]).agg({"output value": cv})



filter_levels = [0, 0.01, 0.02, 0.05, 0.1]

dflist = []
for fl in filter_levels:
    filtered_data = filter_by_quantiles(
                                output,
                                ["short name", "sector", "stage", "sample index"],
                                ["short name", "sector", "stage"],
                                "output value",
                                fl,
                                1-fl)
    
    df_var1 = calculate_variances(filtered_data).reset_index().drop(columns="chunk index").rename(columns={"output value": "variance"})
    df_var2 = pd.DataFrame({
        "variance": np.var(costs, axis=1),
        "short name": [long2short[act["name"]] for act in all_activities],
        "sector": all_sectors
    })
    df_var2["stage"] = "monetization"

    df_var = pd.concat((df_var1, df_var2), ignore_index=True)
    df_var["scenario"] = scen
    df_var["year"] = year
    df_var["location"] = df_var["short name"].apply(lambda x: short2location[x])

    df_sens = normalize_column(
        df_var, "stage", "variance", "sensitivity"
        ).set_index(
            ["short name", "sector", "stage"]
            ).sort_index().reset_index()
    
    df_sens["filter level"] = fl
    dflist.append(df_sens)

df_sens_all = pd.concat(dflist, ignore_index=False)
df_sens_all.to_csv(output_fp+"/gsa_ofat_all_sensdata_w_filters.csv", index=False)

filter_levels = [0, 0.01, 0.02, 0.05, 0.1]

dflist = []
for fl in filter_levels:
    filtered_data = filter_by_quantiles(
                                output,
                                ["short name", "sector", "stage", "sample index"],
                                ["short name", "sector", "stage"],
                                "output value",
                                fl,
                                1-fl)
    
    df_cv1 = calculate_cvs(filtered_data).reset_index().drop(columns="chunk index").rename(columns={"output value": "CV"})
    df_cv2 = pd.DataFrame({
        "CV": np.apply_along_axis(cv, 1, costs),
        "short name": [long2short[act["name"]] for act in all_activities],
        "sector": all_sectors
    })
    df_cv2["stage"] = "monetization"

    df_cv = pd.concat((df_cv1, df_cv2), ignore_index=True)
    df_cv["scenario"] = scen
    df_cv["year"] = year
    df_cv["location"] = df_var["short name"].apply(lambda x: short2location[x])
    
    df_cv["filter level"] = fl
    dflist.append(df_cv)

df_cv_all = pd.concat(dflist, ignore_index=False)
df_cv_all.to_csv(output_fp+"/gsa_ofat_all_cvdata_w_filters.csv", index=False)

def calculate_rolling_variances(df):
    Nsamples = df["sample index"].max() + 1
    
    return df.groupby(["short name", "sector", "stage"]).rolling(Nsamples, min_periods=2).var()

Without quantile filtering

df_var1 = calculate_variances(output).reset_index().drop(columns="chunk index").rename(columns={"output value": "variance"})

df_var2 = pd.DataFrame({
    "variance": np.var(costs, axis=1),
    "short name": [long2short[act["name"]] for act in all_activities],
    "sector": all_sectors
})
df_var2["stage"] = "monetization"

df_var = pd.concat((df_var1, df_var2), ignore_index=True)
df_var["scenario"] = scen
df_var["year"] = year
df_var["location"] = df_var["short name"].apply(lambda x: short2location[x])
df_var.to_csv(output_fp+"/gsa_ofat_all_vardata.csv", index=False)

df_sens = normalize_column(
        df_var, "stage", "variance", "sensitivity"
        ).set_index(
            ["short name", "sector", "stage"]
            ).sort_index().reset_index()

df_sens.to_csv(output_fp+"/gsa_ofat_all_sensdata.csv", index=False)

### Also try standard deviation instead of variance

df_std1 = calculate_stds(output).reset_index().drop(columns="chunk index").rename(columns={"output value": "std"})

df_std2 = pd.DataFrame({
    "std": np.std(costs, axis=1),
    "short name": [long2short[act["name"]] for act in all_activities],
    "sector": all_sectors
})
df_std2["stage"] = "monetization"

df_std = pd.concat((df_std1, df_std2), ignore_index=True)
df_std["scenario"] = scen
df_std["year"] = year
df_std["location"] = df_std["short name"].apply(lambda x: short2location[x])
df_std.to_csv(output_fp+"/gsa_ofat_all_stddata.csv", index=False)

df_sens_std = normalize_column(
        df_std, "stage", "std", "sensitivity"
        ).set_index(
            ["short name", "sector", "stage"]
            ).sort_index().reset_index()

df_sens_std.to_csv(output_fp+"/gsa_ofat_all_sensdata_std.csv", index=False)
